Tagged several revisions:
v2r1 corresponds to CVS rev. 47 (3/15/2013) and includes the ability to change
the default physics list and changed the option parser structure.

v2r2 corresponds to CVS rev. 50 (4/11/2013) and includes the hadron absorber
(the GDML version) that Kevin Y. put in

v2r3 corresponds to CVS rev. 54 (4/17/2013) (the current HEAD version) and
includes changes that Laura made to some of the example scripts